+++ backend/src/main/java/com/ssafy/member/model/dao/MemberRepository.java
@@ -1,6 +1,7 @@
 package com.ssafy.member.model.dao;
 
 import com.ssafy.member.model.dto.Member;
+import com.ssafy.member.model.dto.data.response.MemberLoginResponse;
 import org.apache.ibatis.annotations.Mapper;
 
 
@@ -10,7 +11,7 @@ public interface MemberRepository {
 
    void register(Member member);
    int existsById(String id);
-   Member findById(String id);
+   MemberLoginResponse findById(String id);
    void update(Member member);
    void delete(int memberId);
